Styling changes and adding physical attributes

+++ server/inputs/fighter.inputs.js
@@ -120,11 +120,11 @@ const fighterInputs = `#graphql
         armReach: Float
         legReach: Float
         bodyType: String
-        koPower: Int
-        durability: Int
-        strength: Int
-        endurance: Int
-        agility: Int
+        koPower: Float
+        durability: Float
+        strength: Float
+        endurance: Float
+        agility: Float
     }
 
     input FighterInput {
